Collisions detected!
Bubble can detect collisions & stops when it collides

Code is cleaned up and refactored

(To do: fine tune collision zone. It’s a little weird now, mostly because we’re not centering bubbles on board and my math is an educated guess. We’re also not checking for overlap, only corners - imagine each bubble is placed into a square in which it fits perfectly. Checking is based on the points of the square not the actual circle inside.)